split multiclass files - 1 class per file please!
(The only allowed exception so far is
Exceptions.cs)
